Improve IR and register allocation
The newly two-phase IR replaces the origin one which was coupled with
the the register allocation and made the optimization harder. In
addition, the stack-based virtual machine uses the available registers
inefficiently in RISC-like architecture.

The first-phase IR uses the three-address code (3AC) to simplify the
complex expression. The register allocator could use these information
to generate the more efficient instructions, using the algorithm that
supports the register-base virtual machine. We use the linear-scan
algorithm in this implementation.

The second-phase IR locates the symbol which is used in branch/jump
instruction. Once the amount of the other instructions is confirmed,
the offset/address could be correct.